+++ setup.py
@@ -34,5 +34,5 @@ setup(
         'Operating System :: OS Independent',
     ],
     python_requires='>=3.10',
-    package_dir={'tcat_gtfs_csv_validator': 'tdei_gtfs_csv_validator'},
+    package_dir={'': 'tcat_gtfs_csv_validator'},
 )
