+++ tcat_gtfs_csv_validator/schemas/gtfs_flex/v2.0/booking_rules_schema.csv
