+++ tcat_gtfs_csv_validator/schemas/gtfs_pathways/v1.0/stops_schema.csv
